Finalized app - Cleaning in progress

+++ server/src/controllers/user.image.controller.js
@@ -1,15 +1,13 @@
 /* eslint-disable no-underscore-dangle */
 import fs from "file-system";
-import jwtDecode from "jwt-decode";
 import Image from "../models/image.model";
-import User from "../models/user.model";
-import Course from "../models/courses.model";
 
 const create = async (req, res, next) => {
   const originalFileName = req.file.originalname.split("-")[0];
   const fileName = req.file.originalname;
-  const id = jwtDecode(req.cookies.userJwtToken)._id;
-  const courseId = req.file.originalname.split("_")[2].split("-")[0];
+
+  const imageId = fileName.split("-")[0];
+  console.log(imageId);
 
   // file size threshold
   const fileSizeLimit = 1024 * 1024 * 10;
@@ -41,21 +39,15 @@ const create = async (req, res, next) => {
       // -. Than remove all except last added item
       fs.fs
         .readdirSync("./images")
-        .filter((item) => item.includes(`${originalFileName}`))
+        //.filter((item) => item.includes(`${originalFileName}`))
         .map((item) => {
-          if (item.includes(courseId)) {
-            if (item !== fileName) {
-              fs.fs.unlinkSync(`./images/${item}`);
-            }
-          } else {
-            if (item !== fileName) {
-              fs.fs.unlinkSync(`./images/${item}`);
-            }
+          if (item !== fileName && item.includes(imageId)) {
+            fs.fs.unlinkSync(`./images/${item}`);
           }
         });
 
-      res.send({
-        message: "Cover image uploaded successfully",
+      return res.send({
+        message: "Image uploaded successfully",
         imageUrl: `/images/${fileName}`,
       });
     });
@@ -64,19 +56,6 @@ const create = async (req, res, next) => {
     // Code below unlinks file if there is any error (for example wrong file format)
     fs.fs.unlinkSync(`./${req.file.path}`);
     return res.send({ error: "Format of the file must be PNG | JPEG | JPG" });
-  }
-
-  if (fileName.includes("course")) {
-    const courseId = req.file.originalname.split("_")[2].split("-")[0];
-    const course = await Course.findOneAndUpdate(
-      { _id: courseId },
-      { courseImage: `/images/${fileName}` }
-    );
-  } else {
-    await User.findOneAndUpdate(
-      { _id: id },
-      { userImage: `/images/${fileName}` }
-    );
   }
 };
